+++ domain/src/main/java/com/fintory/domain/stock/dto/websocket/LiveStockPriceStream.java
@@ -2,6 +2,7 @@ package com.fintory.domain.stock.dto.websocket;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 
+import java.io.Serializable;
 import java.math.BigDecimal;
 
 @JsonIgnoreProperties(ignoreUnknown = true)
@@ -10,4 +11,6 @@ public record LiveStockPriceStream(
         BigDecimal currentPrice,
         BigDecimal priceChange,
         BigDecimal priceChangeRate
-){ }
+) implements Serializable {
+    private static final long serialVersionUID = 1L;
+}
